requirements.txt
pandas 
numpy 
scipy 
scikit-learn 
matplotlib 
seaborn
plotly
openai
python-dotenv
langchain 
langchain-openai
flask